requirements.txt

altgraph==0.17
appdirs==1.4.4
asgiref==3.2.7
autopep8==1.5.3
beautifulsoup4==4.9.1
cachetools==4.1.0
camelcase==0.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
comtypes==1.1.7
decorator==4.4.2
dialogflow==1.0.0
distlib==0.3.0
Django==3.0.7
filelock==3.0.12
Flask==1.0.2
Flask-SQLAlchemy==2.4.3
future==0.18.2
geocoder==1.38.1
google-api-core==1.17.0
google-auth==1.14.3
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
lxml==4.5.1
MarkupSafe==1.1.1
MouseInfo==0.1.3
mysqlclient @ file:///C:/Users/gs935/Downloads/mysqlclient-1.4.6-cp38-cp38-win32.whl
numpy==1.18.5
pandas==1.0.5
pefile==2019.4.18
Pillow==7.1.2
prompt-toolkit==3.0.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/gs935/Downloads/PyAudio-0.2.11-cp38-cp38-win32.whl
PyAutoGUI==0.9.50
pycodestyle==2.6.0
pydub==0.24.1
PyGetWindow==0.0.8
PyInstaller==3.6
PyMsgBox==1.0.8
pyperclip==1.8.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
pyttsx3==2.71
PyTweening==1.0.3
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
ratelim==0.1.6
requests==2.23.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.5.0
six==1.14.0
sklearn==0.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
threadpoolctl==2.1.0
toml==0.10.1
urllib3==1.25.9
virtualenv==20.0.20
wcwidth==0.1.9
Werkzeug==1.0.1
wikipedia==1.4.0
wit==5.1.0
youtube-search==0.1.4

beautifulsoup4==4.9.1
bidict==0.19.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
future==0.18.2
geocoder==1.38.1
idna==2.10
jsonlib==1.6.1
loc==0.0.2
PyAudio==0.2.11
pyobjc==6.2.1
pyobjc-core==6.2.1
pyobjc-framework-Accounts==6.2.1
pyobjc-framework-AddressBook==6.2.1
pyobjc-framework-AdSupport==6.2.1
pyobjc-framework-AppleScriptKit==6.2.1
pyobjc-framework-AppleScriptObjC==6.2.1
pyobjc-framework-ApplicationServices==6.2.1
pyobjc-framework-AuthenticationServices==6.2.1
pyobjc-framework-AutomaticAssessmentConfiguration==6.2.1
pyobjc-framework-Automator==6.2.1
pyobjc-framework-AVFoundation==6.2.1
pyobjc-framework-AVKit==6.2.1
pyobjc-framework-BusinessChat==6.2.1
pyobjc-framework-CalendarStore==6.2.1
pyobjc-framework-CFNetwork==6.2.1
pyobjc-framework-CloudKit==6.2.1
pyobjc-framework-Cocoa==6.2.1
pyobjc-framework-Collaboration==6.2.1
pyobjc-framework-ColorSync==6.2.1
pyobjc-framework-Contacts==6.2.1
pyobjc-framework-ContactsUI==6.2.1
pyobjc-framework-CoreAudio==6.2.1
pyobjc-framework-CoreAudioKit==6.2.1
pyobjc-framework-CoreBluetooth==6.2.1
pyobjc-framework-CoreData==6.2.1
pyobjc-framework-CoreHaptics==6.2.1
pyobjc-framework-CoreLocation==6.2.1
pyobjc-framework-CoreMedia==6.2.1
pyobjc-framework-CoreMediaIO==6.2.1
pyobjc-framework-CoreML==6.2.1
pyobjc-framework-CoreMotion==6.2.1
pyobjc-framework-CoreServices==6.2.1
pyobjc-framework-CoreSpotlight==6.2.1
pyobjc-framework-CoreText==6.2.1
pyobjc-framework-CoreWLAN==6.2.1
pyobjc-framework-CryptoTokenKit==6.2.1
pyobjc-framework-DeviceCheck==6.2.1
pyobjc-framework-DictionaryServices==6.2.1
pyobjc-framework-DiscRecording==6.2.1
pyobjc-framework-DiscRecordingUI==6.2.1
pyobjc-framework-DiskArbitration==6.2.1
pyobjc-framework-DVDPlayback==6.2.1
pyobjc-framework-EventKit==6.2.1
pyobjc-framework-ExceptionHandling==6.2.1
pyobjc-framework-ExecutionPolicy==6.2.1
pyobjc-framework-ExternalAccessory==6.2.1
pyobjc-framework-FileProvider==6.2.1
pyobjc-framework-FileProviderUI==6.2.1
pyobjc-framework-FinderSync==6.2.1
pyobjc-framework-FSEvents==6.2.1
pyobjc-framework-GameCenter==6.2.1
pyobjc-framework-GameController==6.2.1
pyobjc-framework-GameKit==6.2.1
pyobjc-framework-GameplayKit==6.2.1
pyobjc-framework-ImageCaptureCore==6.2.1
pyobjc-framework-IMServicePlugIn==6.2.1
pyobjc-framework-InputMethodKit==6.2.1
pyobjc-framework-InstallerPlugins==6.2.1
pyobjc-framework-InstantMessage==6.2.1
pyobjc-framework-Intents==6.2.1
pyobjc-framework-IOSurface==6.2.1
pyobjc-framework-iTunesLibrary==6.2.1
pyobjc-framework-LatentSemanticMapping==6.2.1
pyobjc-framework-LaunchServices==6.2.1
pyobjc-framework-libdispatch==6.2.1
pyobjc-framework-LinkPresentation==6.2.1
pyobjc-framework-LocalAuthentication==6.2.1
pyobjc-framework-MapKit==6.2.1
pyobjc-framework-MediaAccessibility==6.2.1
pyobjc-framework-MediaLibrary==6.2.1
pyobjc-framework-MediaPlayer==6.2.1
pyobjc-framework-MediaToolbox==6.2.1
pyobjc-framework-Metal==6.2.1
pyobjc-framework-MetalKit==6.2.1
pyobjc-framework-ModelIO==6.2.1
pyobjc-framework-MultipeerConnectivity==6.2.1
pyobjc-framework-NaturalLanguage==6.2.1
pyobjc-framework-NetFS==6.2.1
pyobjc-framework-Network==6.2.1
pyobjc-framework-NetworkExtension==6.2.1
pyobjc-framework-NotificationCenter==6.2.1
pyobjc-framework-OpenDirectory==6.2.1
pyobjc-framework-OSAKit==6.2.1
pyobjc-framework-OSLog==6.2.1
pyobjc-framework-PencilKit==6.2.1
pyobjc-framework-Photos==6.2.1
pyobjc-framework-PhotosUI==6.2.1
pyobjc-framework-PreferencePanes==6.2.1
pyobjc-framework-PushKit==6.2.1
pyobjc-framework-Quartz==6.2.1
pyobjc-framework-QuickLookThumbnailing==6.2.1
pyobjc-framework-SafariServices==6.2.1
pyobjc-framework-SceneKit==6.2.1
pyobjc-framework-ScreenSaver==6.2.1
pyobjc-framework-ScriptingBridge==6.2.1
pyobjc-framework-SearchKit==6.2.1
pyobjc-framework-Security==6.2.1
pyobjc-framework-SecurityFoundation==6.2.1
pyobjc-framework-SecurityInterface==6.2.1
pyobjc-framework-ServiceManagement==6.2.1
pyobjc-framework-Social==6.2.1
pyobjc-framework-SoundAnalysis==6.2.1
pyobjc-framework-Speech==6.2.1
pyobjc-framework-SpriteKit==6.2.1
pyobjc-framework-StoreKit==6.2.1
pyobjc-framework-SyncServices==6.2.1
pyobjc-framework-SystemConfiguration==6.2.1
pyobjc-framework-SystemExtensions==6.2.1
pyobjc-framework-UserNotifications==6.2.1
pyobjc-framework-VideoSubscriberAccount==6.2.1
pyobjc-framework-VideoToolbox==6.2.1
pyobjc-framework-Vision==6.2.1
pyobjc-framework-WebKit==6.2.1
pyttsx3==2.88
ratelim==0.1.6
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
urllib3==1.25.9
wikipedia==1.4.0

